meta: updated messaging regarding dates
Currently our schedule specifies exact dates, this can be hard to
commit to. If we plan to be flexible with the dates we chose when
moving a release stream between different levels of support we should
be explicit about it.

This removes all future dates that have currently been decided,
alternatively offering a month in which the event will take place.

It also adds some new language the outlines when individuals should
expect us to give them a date.

 * No later than the first of the month the change is happening
 * No less than 14 days before the change is happening

I believe that this contract is reasonable and much less likely
to cause miscommunication in the future.

One thing not mentioned in here is how we will be communicating these
dates, and potential changes. We need to decide on a single communication
channel and stick to it. Very likely it can be the blog, but we should
discuss this first.